requirements.txt
natsort==7.1.1
pyperclip==1.8.1
webptools==0.0.7
requests==2.26.0
Pillow==8.3.2
PySimpleGUI==4.49.0
